Move handlers into separate folder.
