+++ v3/src/data-interactive/handlers/interactive-frame-handler.ts
@@ -1,5 +1,5 @@
-import { registerDIHandler } from "./data-interactive-handler"
-import { DIHandler, DIResources, diNotImplementedYet, DIValues } from "./data-interactive-types"
+import { registerDIHandler } from "../data-interactive-handler"
+import { DIHandler, DIResources, diNotImplementedYet, DIValues } from "../data-interactive-types"
 
 const noIFResult = {success: false, values: {error: 'Interactive Frame not found'}}
